Added retrofit dependencies and made some refactoring
