+++ app/src/main/java/com/example/movieapp/remote/entites/MoviesDTO.kt
@@ -1,8 +1,8 @@
-package com.example.movieapp.rest_entites
+package com.example.movieapp.remote.entites
 
 import com.example.movieapp.entities.Movie
 
-data class NowPlayingMoviesDTO(
+data class MoviesDTO(
     val page: Int,
     val results: List<Movie>
 )
