+++ app/src/main/res/navigation/mobile_navigation.xml
@@ -6,7 +6,7 @@
     app:startDestination="@+id/navigation_home">
     <fragment
         android:id="@+id/navigation_home"
-        android:name="com.example.movieapp.ui.home.HomeWithServiceFragment"
+        android:name="com.example.movieapp.ui.home.HomeFragment"
         android:label="@string/title_home"
         tools:layout="@layout/fragment_home" >
         <action
